feat: Add external-dns(v0.14.0) as a dependency
Adds external-dns (latest released version v0.14.0) as a go module
dependency and updates the dnsrecord controller and dns provider logic
to align with an external-dns approach to dns management.

dnsrecord changes:
* Replace v1alpha1.Endpoint with externaldns Endpoint
* Add GetRootDomain to DNSRecord, returns the shortest domain that is
  shared across all spec.Endpoints dns names.

provider changes:

* Remove Ensure and Delete and instead include
  externaldnsprovider.Provider in provider.Provider interface.
* Add provider.Config struct to hold common provider configuration,
  currently it only handles filters. Update provider Factory
constructors accept the config as input.
* Set DomainFilter, ZoneTypeFilter and ZoneIDFilter in provider
  implementations from provider.Config.

provider aws changes:

* Implement AdjustEndpoints to change our generic provider specific
  fields for weight and geo into aws ones understood by external-dns.

provider google changes:

* Copy external-dns google provider
* Adds implementations of funtions required to translate endpoints and
  resourceRecordSets into the different states.
  - endpointsToProviderFormat - converts a list of endpoints into a
    google specific format.
  - endpointsFromResourceRecordSets - converts a list of
    `ResourceRecordSet` into endpoints (google format).
  - resourceRecordSetFromEndpoint - converts an endpoint(google format)
    into a `ResourceRecordSet`.

controller changes:

* Update DNSRecord controller to use applyChanges which follows an
  external-dns process to determine a set of
changes(Create/Update/Delete) that are required to be made against the
provider.
* Use noop registry, no owner is currently set or checked for any
  records.
* Setup provider filters:
  - DomainFilter = root domain for endpoints determined by calling
    dnsRecord.GetRootDomain()
  - ZoneTypeFilter = "", public or private zones considered.
  - ZoneIDFilter = set to single hosted zone id (managedZone.Status.ID)
* Use sync policy, all changes will be applied (Create, Update and
  Deletes)
* Use plan with above and using the current zone(plan.Current) and spec
  endpoints(plan.Desired) as input.

additional changes:

downgrade gatewayapiv1 to v0.7.1. This is required since the external
dns aws provider is including the external dns source package which has
a dependency on Gateway API v0.7.1. Downgrading is less than ideal, but
dns-operator shouldn't have a dependency on gateway api, and the health
check code that is currently causing it will likely be removed soon
anyway in favour of provider health checks. There is also an opportunity
to change external-dns to prevent it including the source package since
it's including here for no reason other than getting a version
https://github.com/kubernetes-sigs/external-dns/blob/master/provider/aws/session.go#L72