registry: Use txt registry for owned records
Use the TXT Registry when an `ownerID` is specified for a record being
processed.  Adds an optional field `ownerID` to the DNSRecord spec. This
is immutable since changing it after initial creation will likely not
result in desirable behaviour.

Update e2e tests to verify contents of zone after record creation.