Import external dns plan package and txt registry
Move the external-dns plan package into this repo along with the txt
registry.

To ease development we will initially just work off a copy of the
relevant external-dns code in this repo. When we have a solution that
works for us we will look into how we can submit that back to
external-dns.

Code is copied, unmodified where possible, from the v0.14.1 version of
external-dns. https://github.com/kubernetes-sigs/external-dns/tree/v0.14.1