Add RootHost and CNAME target validations
Add step in the plan to verify the targets of all CNAME records being
created or updated. If the target is calculated to be a subdomain of the
plans root host, it must exist in the zone or will exist after the
changes are applied.

Makes use of the plans DomainFilters since this is already used to
filter down to relevant records and allows us to use the same logic to
check for matches in dnsName values or targets. Plan was updated to
ensure that a DomainFilter is always set for the RootHost if one is
specified.