Make ownerID and rootHost required
Updates the DNSRecord spec to make `ownerID` and `rootHost` required. We
might in the future decide to change this, but for now this is the most
tested path and what is required for the kuadrant DNSPolicy. This will
also prevent a situation where a DNSRecord could be created that can
unexpectedly remove unrelated records from a zone.

Adds minimal validation to the fields, length > 1 for both `ownerID` and
`rootHost` and a max length of 12 for `ownerID`.