tests: Add provider errors e2e specs
Adds a provider errors set of specs to the e2e test suite. The purpose
of this is to test each supported provider returns a reasonable response
and updates the record status appropriately in cases where a user can
misconfigure a DNSRecord and cause a provider error. These mainly focus
on what can be set from the DNSPolicy (geo and weight) and the intention
isn't to exhaustively test all error scenarios.

Added as an e2e test since this can, and will be, different on every
provider. Tests also ensure that the records are not falling into a loop
of constantly being updated when a provider error is updating the
status.