external-dns: Remove logrus logger
Removes the use of logrus in all external-dns code in favour of setting
a `logr.Logger` on all structs.

Constructors are used throughout to create new instances of components
(plan, registry and providers), with each requiring the current context
and a logger instance retrieved from it with an appropriate logger name
added where required.