feat: Generate unique OwnerID based on DNSRecord UID
Makes the ownerID in the DNSrecord spec optional. If no ownerID is set
in the DNSRecord spec on initial creation, the dns record controller
itself will generate a unique ownerID value based on the DNSRecord UID.
The spec ownerID is still immutable, but now also has additional
validations to ensure it can never be changed from a set to unset or
unset to set state after initial creation.
Since the spec ownerID is now optional, the ownerID used throughout all
operations is stored and retrieved from the status of the DNSRecord.

Adds additional validations to the ownerID field specifying a MinLength
and MaxLength as per best practices for CEL validations. MaxLength is
currently 36 which allows the full UID value to be used if desirable.