chore: Immutable RootHost field on DNSRecord
The dnsrecord controller does not currently handle changes to the
rootHost so is now marked as immutable to prevent it from being changed
and causing odd behaviour.

In the case where you want to change the host, a new record should be
created and the old one removed to ensure proper cleanup of resources in
the provider.

Signed-off-by: Michael Nairn <mnairn@redhat.com>