bug: Sanitize provider error messages on zone lookup
Fixes an issue where if invalid credentials are used in the provider
secret when looking for a suitable zone in the provider it would
continually change the status on every reconcile. This is caused by the
AWS provider adding a request id to the error which wasn't being removed
in this case prior to update the record status.

Signed-off-by: Michael Nairn <mnairn@redhat.com>